Create tokenvestingservice, fix styles & refactor
