+++ src/components/UserPage/styled/index.ts
@@ -7,9 +7,9 @@ export const TitleRow = styled.div`
 
 export const ListRow = styled.div`
   display: flex;
-  alignitems: center;
+  align-items: center;
   padding: 6px 0px;
-  borderbottom: ${({ borderBottom }) => (borderBottom ? '1px solid' : '')};
+  border-bottom: ${({ borderBottom }) => (borderBottom ? '1px solid' : '')};
   cursor: ${({ clickable }) => (clickable ? 'pointer' : 'default')};
   &:hover {
     opacity: ${({ clickable }) => (clickable ? '0.5' : '1')};
