+++ src/contexts/index.ts
@@ -13,6 +13,7 @@ import PoktService from '../services/PoktService';
 import AlchemyService from '../services/AlchemyService';
 import CustomRpcService from '../services/CustomRpcService';
 import ENSService from '../services/ENSService';
+import TokenVestingService from '../services/TokenVestingService';
 
 import ProviderStore from '../stores/Provider';
 import TransactionStore from '../stores/Transaction';
@@ -59,6 +60,7 @@ export default class RootContext {
   alchemyService: AlchemyService;
   customRpcService: CustomRpcService;
   ensService: ENSService;
+  tokenVestingService: TokenVestingService;
 
   constructor() {
     this.providerStore = new ProviderStore(this);
@@ -83,6 +85,7 @@ export default class RootContext {
     this.alchemyService = new AlchemyService(this);
     this.customRpcService = new CustomRpcService(this);
     this.ensService = new ENSService(this);
+    this.tokenVestingService = new TokenVestingService(this);
   }
 
   reset() {
