+++ src/services/DaoService.ts
@@ -4,7 +4,6 @@ import PromiEvent from 'promievent';
 import RootContext from '../contexts';
 import { ContractType } from '../stores/Provider';
 import { BigNumber, MAX_UINT } from '../utils';
-import { encodeDxdVestingRelease } from '../utils/encodingCalls';
 
 export default class DaoService {
   context: RootContext;
@@ -355,19 +354,6 @@ export default class DaoService {
         },
         [proposalId]
       ),
-      '0'
-    );
-  }
-
-  redeemVestingContractDxd(contractAddress: string): PromiEvent<any> {
-    const { providerStore, configStore } = this.context;
-    const { library } = providerStore.getActiveWeb3React();
-    const contracts = configStore.getNetworkContracts();
-
-    return providerStore.sendRawTransaction(
-      providerStore.getActiveWeb3React(),
-      contractAddress,
-      encodeDxdVestingRelease(library, contracts.votingMachines.dxd.token),
       '0'
     );
   }
